+++ atst/domain/csp/__init__.py
@@ -1,5 +1,3 @@
-import importlib
-
 from .cloud import MockCloudProvider
 from .file_uploads import AzureUploader, MockUploader
 from .reports import MockReportingProvider
@@ -31,20 +29,3 @@ def make_csp_provider(app, csp=None):
         app.csp = MockCSP(app, test_mode=True)
     else:
         app.csp = MockCSP(app)
-
-
-def _stage_to_classname(stage):
-    return "".join(map(lambda word: word.capitalize(), stage.split("_")))
-
-
-def get_stage_csp_class(stage, class_type):
-    """
-    given a stage name and class_type return the class
-    class_type is either 'payload' or 'result'
-
-    """
-    cls_name = f"{_stage_to_classname(stage)}CSP{class_type.capitalize()}"
-    try:
-        return getattr(importlib.import_module("atst.domain.csp.cloud"), cls_name)
-    except AttributeError:
-        print("could not import CSP Result class <%s>" % cls_name)
