+++ atst/domain/csp/cloud/cloud_provider_interface.py
@@ -124,3 +124,4 @@ class CloudProviderInterface:
         exception if an error occurs while creating a subscription.
         """
         raise NotImplementedError()
+
